app: simplify USB DFU Kconfig options
Simplify the USB DFU Kconfig options. Instead of duplicating these between
the app and sysbuild, have sysbuild read them from the app and apply them
to MCUboot.

Add a top-level Kconfig option for enabling DFU support and a choice for
selecting the backend. Two backends are supported as of now, MCUboot and
None - the latter is meant for use with on-chip ROM-based DFU bootloaders.

Signed-off-by: Henrik Brix Andersen <henrik@brixandersen.dk>